refactor: move referencegrant agent/internal → common (026 EM1c prep)
Pure ReferenceGrant logic (PermitsBackend/CrossNamespace) used by the GAMMA + L4
+ edge reconcilers. Moving it to common/ lets the upcoming shared GAMMA projector
(026 EM1c — used by BOTH the agent reconciler and the registrar export controller)
reuse it without importing agent internals. No behaviour change; importers updated.
