+++ agent/internal/edge/gatewayapi/BUILD.bazel
@@ -15,12 +15,12 @@ go_library(
         "//agent/internal/edge/portalloc",
         "//agent/internal/edge/secret",
         "//agent/internal/gatewaystatus",
-        "//agent/internal/referencegrant",
         "//agent/internal/xds/cache",
         "//agent/internal/xds/proxy",
         "//api/aether/config/v1:config",
         "//common/constants",
         "//common/log",
+        "//common/referencegrant",
         "//common/serviceref",
         "@io_k8s_api//core/v1:core",
         "@io_k8s_apimachinery//pkg/api/errors",
