+++ agent/internal/edge/gatewayapi/status.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/bpalermo/aether/agent/internal/gatewaystatus"
-	"github.com/bpalermo/aether/agent/internal/referencegrant"
+	"github.com/bpalermo/aether/common/referencegrant"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
